Reorganize shapes export
