+++ src/shapes/index.js
@@ -1,16 +1,2 @@
-import PropTypes from 'prop-types'
-
-export const connectorShape = PropTypes.shape({
-  subscribe: PropTypes.func.isRequired,
-  unsubscribe: PropTypes.func.isRequired,
-  connection: PropTypes.shape({
-    bind: PropTypes.func.isRequired,
-    unbind: PropTypes.func.isRequired,
-  }).isRequired,
-})
-
-export const channelShape = PropTypes.shape({
-  bind: PropTypes.func.isRequired,
-  unbind: PropTypes.func.isRequired,
-  trigger: PropTypes.func,
-})
+export * from './channelShape'
+export * from './conectorShape'
